requirements.txt
adium-theme-ubuntu==0.3.4
attrs==16.0.0
cffi==1.7.0
cryptography==1.4
cssselect==0.9.2
decorator==4.0.6
Django==2.0.1
djangorestframework==3.8.2
dnspython==1.12.0
enum34==1.1.6
idna==2.1
ipaddress==1.0.16
ipython==2.4.1
lxml==3.6.1
parsel==1.0.2
pbr==1.9.1
pexpect==4.0.1
ptyprocess==0.5
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pycrypto==2.6.1
PyDispatcher==2.0.5
pyOpenSSL==16.0.0
queuelib==1.4.2
Scrapy==1.1.1
service-identity==16.0.0
simplegeneric==0.8.1
six==1.10.0
stevedore==1.13.0
Twisted==16.3.0
unity-lens-photos==1.0
virtualenv==15.0.1
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
w3lib==1.14.2
WebOb==1.5.1
zope.interface==4.2.0
